// apps/client/src/generatedClient/models/UserIdentifier.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserIdentifier = ({
    id: number;
    email?: string;
    username?: string;
} | {
    id?: number;
    email: string;
    username?: string;
} | {
    id?: number;
    email?: string;
    username: string;
});